requirements.txt
streamlit
openai==0.28.1
Pillow
requests